depinstc: small speed improvement
By default, the namespace method is used for user-chroot, but when a
port fails, the process of installing in the container is retried with
the proot method. This order is in place because the namespace method
is faster, but some ports may fail.

This patch makes sure the proot method doesn't stick for all the following
ports, and let them build with the namespace method first. It also take
of the fact that depinstc could be started with FUNC_PROOT=1 (to force
the proot method first). The namespace method won't be used, because
there is no chance a build could fail with proot but not with namespaces.